chore: update router file to TS
